+++ src/models/regional/audit_logs.ts
@@ -1,0 +1,6 @@
+/** @see https://developers.scalingo.com/audit_logs */
+export interface AuditLog {
+  time: string;
+  content: string;
+  type: string;
+}